% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLincsCor-pkg.R
\docType{package}
\name{iLincsCor-package}
\alias{iLincsCor}
\alias{iLincsCor-package}
\title{iLincsCor: Fast signature correlations}
\description{
Load signature libraries to memory to perform fast correlation with one or multiple inputs.
}
\details{
iLincsCor is a package containing iLincsCor class for fast signature 
correlations.
}
